Deploy/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow
streamlit
requests
pandas
